Use ZoomLevel for viewport and ride window

+++ src/openrct2/interface/Viewport.cpp
@@ -177,8 +177,8 @@ void viewport_create(rct_window* w, const ScreenCoordsXY& screenCoords, int32_t 
     viewport->height = height;
     const auto zoom = focus.zoom;
 
-    viewport->view_width = width << zoom;
-    viewport->view_height = height << zoom;
+    viewport->view_width = width << static_cast<int8_t>(zoom);
+    viewport->view_height = height << static_cast<int8_t>(zoom);
     viewport->zoom = zoom;
     viewport->flags = 0;
